pageserver: remove un-needed "uninit mark"
Previously, existence of a timeline directory was treated
as evidence of the timeline's logical existence.  That is no
longer the case since we treat remote storage as the source
of truth on each startup: we can therefore do without
this mark file.

Deletion of the mark files in Tenant::clean_up_timelines
is retained temporarily: after a couple of releases that
can go away too.